+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+## [0.0.5] - 2016-12-28
 ### Added
 - Implemented the "nomargin" property on pf-layout.
 - Implemented v-tooltip directive using tether-tooltip.
@@ -62,7 +64,8 @@ disabled, as per PatternFly design guidelines.
 - pf-toolbar component
 - pf-utilization-bar-chart component
 
-[Unreleased]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.4...HEAD
+[Unreleased]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.5...HEAD
+[0.0.5]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.1...v0.0.2